pyproject.toml
[project]
name = "ds-vpr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "accelerate>=1.11.0",
    "datasets>=4.3.0",
    "diffusers>=0.35.2",
    "hf-transfer>=0.1.9",
    "ipykernel>=7.1.0",
    "openai>=2.8.1",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "torch>=2.9.0",
    "transformers>=4.57.1",
]
